Re-factor getValidatedConfig to getConfig and validateConfig. Fix #19
