+++ lib/fliff-cli.js
@@ -4,13 +4,13 @@ import * as path from 'path';
 import pjson from '../package.json';
 import './colors-set-theme';
 import { FLIFF } from './fliff.js';
-import { commandErrorHandler, getValidatedConfig } from './shared';
+import { commandErrorHandler, getConfig, validateConfig } from './shared';
 import fliffUsage, { addUsage, updateUsage, getUsage, deleteUsage, initUsage, configUsage, tokenUsage } from './fliff-cli-usage';
 
 try {
     const commandLineArgs = require('command-line-args');
     const commandLineUsage = require('command-line-usage');
-    
+
     const { operation, _unknown } = commandLineArgs([
         { name: 'operation', defaultOption: true }
     ], { stopAtFirstUnknown: true });
@@ -31,8 +31,8 @@ try {
         { name: 'help', alias: 'h', type: Boolean }
     ], { argv });
     const fliff = new FLIFF();
-    
-    
+
+
     if (options.help) {
         switch (operation) {
             case 'add': console.log(commandLineUsage(addUsage)); break;
@@ -46,12 +46,13 @@ try {
         }
         process.exit(0);
     }
-    
+
     // Commands that need Functions config
     if (['add', 'update', 'delete', 'get', 'token'].indexOf(operation) > -1) {
-        getValidatedConfig().then(() => {
+        getConfig().then(config => {
             switch (operation) {
                 case 'add':
+                    validateConfig(config);
                     console.log('Sending request to add LIFF view...'.verbose);
                     fliff.add(options)
                         .then(rsAdd => {
@@ -63,10 +64,11 @@ try {
                             console.log(message.error);
                             process.exit(1);
                         });
-    
+
                     break;
-    
+
                 case 'delete':
+                    validateConfig(config);
                     console.log(`Sending request to delete LIFF view(s)...`.verbose);
                     fliff.delete(options)
                         .then(rsDelete => {
@@ -80,8 +82,9 @@ try {
                             process.exit(1);
                         });
                     break;
-    
+
                 case 'get':
+                    validateConfig(config);
                     console.log('Sending request to get LIFF view(s)...'.verbose);
                     fliff.get(options)
                         .then(rsGet => {
@@ -97,15 +100,16 @@ try {
                             console.log(message.error);
                             process.exit(1);
                         });
-    
+
                     break;
-    
+
                 case 'update':
+                    validateConfig(config);
                     console.log(`Sending request to update LIFF view`.verbose);
                     fliff.update(options)
                         .then(rsUpdate => {
                             console.log(`Updated LIFF ID: ${options.id.input}`.info);
-    
+
                             return rsUpdate;
                         })
                         .catch(errUpdate => {
@@ -114,20 +118,20 @@ try {
                             process.exit(1);
                         });
                     break;
-    
+
                 case 'token':
                     if (options.issue === true) {
                         console.log('Issuing channel access token'.verbose);
                     } else if (options.revoke !== undefined) {
                         console.log('Revoking channel access token'.verbose);
                     }
-    
+
                     fliff.token(options)
                         .then(rsToken => {
                             if (options.issue === true && rsToken.accessToken) {
                                 console.log(`The following token has been issued.`.info);
                                 console.log(JSON.stringify(rsToken, undefined, 2));
-    
+
                                 if (options.save === true) {
                                     console.log(`The access token is saved on Firebase Functions Configuration.`.info);
                                 } else {
@@ -139,7 +143,7 @@ try {
                             } else {
                                 console.log('Unknown response').warn;
                             }
-    
+
                             return rsToken;
                         })
                         .catch(errToken => {
@@ -147,17 +151,17 @@ try {
                             console.log(message.error);
                             process.exit(1);
                         });
-    
+
                     break;
-    
+
                 default:
-    
+
             }
-    
+
         });
     } else if (operation) {
         const initPath = path.resolve(process.cwd(), 'web-views');
-    
+
         switch (operation) {
             case 'init':
                 fliff.init(initPath)
@@ -186,7 +190,7 @@ try {
                         console.log(`The following property has been updated.`.info);
                         console.log(JSON.stringify(rsConfig, undefined, 2));
                         console.log(`Firebase Functions configured`.info);
-    
+
                         return rsConfig;
                     })
                     .catch(errConfig => {
@@ -203,7 +207,7 @@ try {
         console.log(`Version: ${pjson.version}`);
     } else {
         console.log(commandLineUsage(fliffUsage));
-    }    
-} catch(commandError) {
+    }
+} catch (commandError) {
     commandErrorHandler(commandError);
 }
