+++ lib/richmenu-cli.js
@@ -6,7 +6,7 @@ import * as util from 'util';
 import './colors-set-theme';
 import pjson from '../package.json';
 import { LIFFConfig, RichMenuAddRequest, RichMenuDeleteRequest, RichMenuLinkUserRequest, RichMenuGetRequest, RichMenuUploadRequest, RichMenuSetDefaultRequest, RichMenuUnlinkUserRequest } from './index';
-import { commandErrorHandler, getValidatedConfig } from './shared';
+import { commandErrorHandler, getConfig, validateConfig } from './shared';
 
 try {
     const commandLineArgs = require('command-line-args');
@@ -28,7 +28,7 @@ try {
 
     // Commands that need Functions config
     if (['add', 'delete', 'get', 'default', 'link', 'unlink'].indexOf(operation) > -1) {
-        getValidatedConfig().then(async (config) => {
+        getConfig().then(async (config) => {
             let accessToken = config.line.access_token;
             let names;
             let data;
@@ -37,6 +37,8 @@ try {
             let req;
             let res;
             let uploadReq;
+
+            validateConfig(config);
 
             switch (operation) {
                 case 'add':
